Added basic draggable ingredient functionality
Created a draggable ingredient prefab, and a corresponding pot prefab to implement basic drag-and-drop functionality to the soup scene. NOTE: this is only a proof of concept for the drag and drop feature - this is NOT YET a functional replacement for the button system.

TODO: finish adding functionality to properly replace the old button system; create prefab variants for all ingredients; add the new "pop up info" UI