Draggable ingredients have fully replaced buttons
The old button-based ingredient system has been replaced with the drag-&-drop system.
KNOWN ISSUE: due to a bug with how Unity calculates the hitbox of the "make soup" button, clicking pretty much anywhere onscreen will activate the button, which can cause a scene transition the second 3 ingredients are added to the pot. The cause of this issue is known and will (hopefully) be fixed in the next update.