Resolution options now only refresh on apply
Resolution options list is now repopulated only when the "apply" button is clicked, instead of whenever the display mode selection is changed (but not necessarily applied). Should result in slightly more consistent behavior.

KNOWN BUG: if set to a resolution other than the native one, the resolution selection will "oscillate" between the non-native and native resolutions every time the apply button is pressed (i.e. clicking apply sets the game's resolution to A and the selection to B, clicking again sets the resolution to B and selection to A, and so on).