Partial bugfix of incorrect resolution selection
Partial fix to bug in last commit (found to be caused by resolution list refresh executing too soon and reading an outdated resolution as the current one). Works in fullscreen mode; windowed mode needs more work (see below).

KNOWN ISSUE: changing to windowed mode defaults the resolution selection to native resolution, although it won't apply the change. This is due to the special behavior of Screen.currentResolution in windowed mode; a workaround is in progress.